Change breakToken based on css class
by search for an element with the `avoid-split` class, it is possible to
guide Paged.js on where it should break pages. What should happen now is
when Paged.js tries to split a page, it will check if the breakToken is
inside an element with the `avoid-split` class. If it is, then the
breakToken will get moved to that element instead.