Use `flex: row` for finding sections
The previous method of using absolute/relative positioning and then
inserting the section title with css was causing consistency issues.
This included problems with paged.js deciding that incorrect sections
were the location of a breakToken, which resulted in some wonkyness.

The finding table sections now use `flex: row` and the section header is
inserted with javascript inside the paged.js `afterRendered` hook. the
finding-content html in CVSSFinging.html now contains two divs, one for
the section header and another for the content body. If a finding
section needs to be split by paged.js, the section header will be
reinserted onto the next page as intended.