Use table instead of divs for FindingsList
Due to the fact that paged.js handles splitting tables fairly well by
default, the FindingsList component was updated to use table elements
instead of divs with flex. Flex row was also producing similar problems
as flex column on closer inspection, so the consistency that tables
should bring is welcome.

In order to make the tables render with the correct size on the page, a
new container class was created in globals.css that defines some css
variables for page width/height and is set to inherit the width/height
of the pagedjs_page_content. As the container element is a direct child
of the pagedjs_page_content element, values for height/width should work
as expected.