Remove empty finding-content in afterRender function
We remove the empty findings from the DOM only after we have made sure
to transfer the finding-content-headers to the next page as needed. In
some scenarios the finding-content gets split in such a way that all of
the content body is on the second page. In this case, the empty finding
content will get deleted along with the finding-content-header, which
then means we cannot add it back to the second page.