Use highlight.js instead of pygments
Due to the fact that we have to work around paged.js for the reports,
highlight.js preserves white space better and in more cases across page
splits than pygments does. This also removes pygments as a pip
dependency. For now, the Monokai colorscheme is used from the highlight.js
website