Preserve first indent in <pre><code> sections
When splitting "<pre><code>" across pages, the indentation of the first
line on the next page would be lost. This fix is copy pasted from a
paged.js issue where instead of checking if the parent node is PRE, we
instead check if it has pre inside of the css white-space rule.