fix finding table again, please work
using "display: table-cell" also has some problems and just straight up
does not work in firefox. We are going back to old school classics and
using float with a margin. Since the finding-content-headers are always
11rem wide, they can jsut be floated left and the content can be given a
margin-left of 11rem as well. Flex column is used to make the whole
finding table stack correctly, and as long as flex is used in simple
cases, it does not seem to make paged.js wanna die.

Having tried out rendering a report on firefox, chrome, and with the pdf
render endpoint, they all seem to render exactly the same. Adding page
breaks manually also did not break anything.