Fix issue with removing empty finding-content-body
Scenarios where a page break is inserted at the begining of the
finding-content-body would not result in the element getting deleted
afterwards. Since the finding-content-body sticks around if it has a
page break in it, we instead check if the finding-content-body text is
empty. If the finding-content-body only has whitespace, it will be
considered empty.