Change the shape of report.revisions return value
report.revisions used to return a dictionary where the key is the
component id and the value is a list of revisions. In order to add the
name of the component in the result as well, it now returns a list of
dicts that contain the component id, name, and revisions.

Example:
[
    {
        "id": <uuid>,
        "name": <component name>,
        "revisions": [
            <revision 1>
            <revision 2>
            ...
        ]
    },
    ...
]