Create two seperate methods for getting report revisions
The method report.component_revisions() will just get revisions for the
report components and will shape the data so that the revisions are
grouped by the component it revises.

The revisions property on components will return a list of all
revisions, including revisions on findings, in a list for displaying on
the frontend as a changelog.

In order to keep the previous functionality of toggling the display of
latest revisions on the report edit page, it was simpler to split the
functionality into two separate method calls.