Do not remove .finding-content in afterPageLayout()
Due to the logic in the afterRendered function, it is necessary to keep
any dangling elements with the "finding-content" class. Otherwise, there
will be no way for the afterRender function to copy
"finding-content-header" elements as needed. The afterRender function
already had logic to cleanup empty "finding-content" elements, so
removing the offending line in afterPageLayout should be unlikely to
cause issues