Remove padding/margin at beginning/end of page
Due to the way paged.js splits pages, any extra margin/padding will
generally cause an overflow into the footer of the page. Attempts have
been made to work around this, but now we will just work with it. This
should make the page content for findings much more consistent with
sizing.

Additionally, the way code blocks are handled at the end/beginning of a
finding table has been altered slightly to look better without the
padding it once had.