Set overflow-wrap: anywhere in td elements
Setting overflow-wrap: anywhere avoids a very annoying issue with tables
where a table row at the end of a page cold result in a page break
within the row if the contents overflowed. This results in the custom
paged.js hooks prepending the table header onto the original table,
causing the table to have two headers.

This is not a perfect fix, and paged.js will still sometimes try to
split table rows if it absolutely necessary. A better fix may be
required in the future, but for now this seems to alleviate the issue.