Fix issues with DREAD tables on page break
In some cases, a DREAD table would have an empty finding-content-header
html element copied over to the next page on page break. This would push
the header on the next page into the body of the finding-content. This
change just deletes these empty elements from the DOM to ensure that
this does not happen.